requirements.txt
flask
minio
Pillow
